+++ test/c/main.cpp
@@ -10,6 +10,7 @@
 #include "nested.json.c.h"
 #include "repeated.json.c.h"
 #include "oneof.json.c.h"
+#include "optional.json.c.h"
 #include "importing.json.c.h"
 #include "bytes.json.c.h"
 #include "size.json.c.h"
@@ -211,6 +212,60 @@ void test_oneof() {
   oneof_Test_destroy(&b);
 }
 
+void test_optional() {
+  optional_Test a;
+  optional_Test_init(&a);
+  optional_Test b;
+  optional_Test_init(&b);
+  assert(a._a_case == optional_Test_ACase_NOT_SET);
+  TEST_IDENTIFY(optional_Test, &a, &b);
+  assert(b._a_case == optional_Test_ACase_NOT_SET);
+
+  optional_Test_set_a(&a, 1);
+  assert(a._a_case == optional_Test_ACase_kA);
+  assert(a.a == 1);
+  TEST_IDENTIFY(optional_Test, &a, &b);
+  assert(b._a_case == optional_Test_ACase_kA);
+  assert(b.a == 1);
+
+  optional_Test_set_b(&a, "foo");
+  assert(a._b_case == optional_Test_BCase_kB);
+  assert(a.b_len == 3 && strcmp(a.b, "foo") == 0);
+  TEST_IDENTIFY(optional_Test, &a, &b);
+  assert(b._b_case == optional_Test_BCase_kB);
+  assert(b.b_len == 3 && strcmp(b.b, "foo") == 0);
+
+  optional_Test_set_c(&a, optional_BAR);
+  assert(a._c_case == optional_Test_CCase_kC);
+  assert(a.c == optional_BAR);
+  TEST_IDENTIFY(optional_Test, &a, &b);
+  assert(b._c_case == optional_Test_CCase_kC);
+  assert(b.c == optional_BAR);
+
+  optional_Message m;
+  optional_Message_init(&m);
+  optional_Message_set_name(&m, "bar");
+  optional_Test_set_d(&a, &m);
+  optional_Message_destroy(&m);
+  assert(a._d_case == optional_Test_DCase_kD);
+  assert(a.d.name_len == 3 && strcmp(a.d.name, "bar") == 0);
+  TEST_IDENTIFY(optional_Test, &a, &b);
+  assert(b._d_case == optional_Test_DCase_kD);
+  assert(b.d.name_len == 3 && strcmp(b.d.name, "bar") == 0);
+
+  optional_Test_clear__a_case(&a);
+  assert(a._a_case == optional_Test_ACase_NOT_SET);
+  optional_Test_clear__b_case(&a);
+  assert(a._b_case == optional_Test_BCase_NOT_SET);
+  optional_Test_clear__c_case(&a);
+  assert(a._c_case == optional_Test_CCase_NOT_SET);
+  optional_Test_clear__d_case(&a);
+  assert(a._d_case == optional_Test_DCase_NOT_SET);
+
+  optional_Test_destroy(&a);
+  optional_Test_destroy(&b);
+}
+
 void test_importing() {
   importing_Test a;
   importing_Test_init(&a);
@@ -257,6 +312,7 @@ int main() {
   test_nested();
   test_repeated();
   test_oneof();
+  test_optional();
   test_importing();
   test_bytes();
   test_size();