+++ test/unity/JsonifUnityTest/Assets/Test.cs
@@ -97,9 +97,9 @@ public class Test : MonoBehaviour
     void TestOneof()
     {
         var a = new Oneof.Test();
-        D.Assert(a.test_oneof_case == Oneof.Test.TestOneofCase.TEST_ONEOF_NOT_SET);
+        D.Assert(a.test_oneof_case == Oneof.Test.TestOneofCase.NOT_SET);
         a = Identify(a);
-        D.Assert(a.test_oneof_case == Oneof.Test.TestOneofCase.TEST_ONEOF_NOT_SET);
+        D.Assert(a.test_oneof_case == Oneof.Test.TestOneofCase.NOT_SET);
 
         a.SetA(1);
         D.Assert(a.test_oneof_case == Oneof.Test.TestOneofCase.kA);
@@ -130,7 +130,58 @@ public class Test : MonoBehaviour
         D.Assert(a.d.name == "bar");
 
         a.ClearTestOneofCase();
-        D.Assert(a.test_oneof_case == Oneof.Test.TestOneofCase.TEST_ONEOF_NOT_SET);
+        D.Assert(a.test_oneof_case == Oneof.Test.TestOneofCase.NOT_SET);
+    }
+
+    void TestOptional()
+    {
+        var a = new Optional.Test();
+        D.Assert(a._a_case == Optional.Test.ACase.NOT_SET);
+        D.Assert(a._b_case == Optional.Test.BCase.NOT_SET);
+        D.Assert(a._c_case == Optional.Test.CCase.NOT_SET);
+        D.Assert(a._d_case == Optional.Test.DCase.NOT_SET);
+        a = Identify(a);
+        D.Assert(a._a_case == Optional.Test.ACase.NOT_SET);
+        D.Assert(a._b_case == Optional.Test.BCase.NOT_SET);
+        D.Assert(a._c_case == Optional.Test.CCase.NOT_SET);
+        D.Assert(a._d_case == Optional.Test.DCase.NOT_SET);
+
+        a.SetA(1);
+        D.Assert(a._a_case == Optional.Test.ACase.kA);
+        D.Assert(a.a == 1);
+        a = Identify(a);
+        D.Assert(a._a_case == Optional.Test.ACase.kA);
+        D.Assert(a.a == 1);
+
+        a.SetB("foo");
+        D.Assert(a._b_case == Optional.Test.BCase.kB);
+        D.Assert(a.b == "foo");
+        a = Identify(a);
+        D.Assert(a._b_case == Optional.Test.BCase.kB);
+        D.Assert(a.b == "foo");
+
+        a.SetC(Optional.Enum.BAR);
+        D.Assert(a._c_case == Optional.Test.CCase.kC);
+        D.Assert(a.c == Optional.Enum.BAR);
+        a = Identify(a);
+        D.Assert(a._c_case == Optional.Test.CCase.kC);
+        D.Assert(a.c == Optional.Enum.BAR);
+
+        a.SetD(new Optional.Message() { name = "bar" });
+        D.Assert(a._d_case == Optional.Test.DCase.kD);
+        D.Assert(a.d.name == "bar");
+        a = Identify(a);
+        D.Assert(a._d_case == Optional.Test.DCase.kD);
+        D.Assert(a.d.name == "bar");
+
+        a.ClearACase();
+        D.Assert(a._a_case == Optional.Test.ACase.NOT_SET);
+        a.ClearBCase();
+        D.Assert(a._b_case == Optional.Test.BCase.NOT_SET);
+        a.ClearCCase();
+        D.Assert(a._c_case == Optional.Test.CCase.NOT_SET);
+        a.ClearDCase();
+        D.Assert(a._d_case == Optional.Test.DCase.NOT_SET);
     }
 
     void TestImporting()